# api/openapi.yaml
openapi: 3.0.3
info:
  title: CareDesk API
  version: 1.0.0
  description: |
    REST API for the CareDesk platform.

    All resources are served under the `/api/v1` base path. Error responses follow
    RFC 9457 (Problem Details for HTTP APIs) and are returned with the
    `application/problem+json` media type.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api/v1
    description: Versioned base path
security:
  - bearerAuth: []
...
paths:
...
  /patients/{patientId}:
    get:
      tags: [Patients]
      summary: Get patient profile
      operationId: getPatientById
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Patient profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /patients/{patientId}/appointments:
    get:
      tags: [Patients]
      summary: List a patient's appointments (paginated)
      operationId: listPatientAppointments
      parameters:
        - $ref: '#/components/parameters/PatientId'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated patient appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedAppointmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /patients/{patientId}/visit-history:
    get:
      tags: [Patients]
      summary: View a patient's visit history
      operationId: getPatientVisitHistory
      parameters:
        - $ref: '#/components/parameters/PatientId'
      responses:
        '200':
          description: Visit history summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    PatientId:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      description: Zero-based page index.
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeParam:
      name: size
      in: query
      required: false
      description: Number of items per page.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  responses:
    BadRequest:
      description: The request is malformed or fails validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Unauthorized:
      description: Authentication is required or has failed.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    Forbidden:
      description: The caller is authenticated but not allowed to access the resource.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InternalServerError:
      description: An unexpected error occurred while processing the request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
    ProblemDetail:
      type: object
      description: |
        Problem Details object as defined by RFC 9457
        (https://www.rfc-editor.org/rfc/rfc9457).
      required: [title, status]
      properties:
        type:
          type: string
          format: uri
          description: A URI reference identifying the problem type.
          default: about:blank
        title:
          type: string
          description: A short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          description: The HTTP status code generated by the origin server.
          minimum: 100
          maximum: 599
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
        instance:
          type: string
          format: uri
          description: A URI reference that identifies the specific occurrence.
        errors:
          type: array
          description: Optional list of field-level validation errors.
          items:
            $ref: '#/components/schemas/ValidationError'
    ValidationError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
    PageMeta:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page:
          type: integer
          minimum: 0
          description: Zero-based page index of the returned slice.
        size:
          type: integer
          minimum: 1
          description: Number of items per page.
        totalElements:
          type: integer
          format: int64
          minimum: 0
          description: Total number of items across all pages.
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages.
...
    PaginatedAppointmentResponse:
      type: object
      required: [content, page]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
        page:
          $ref: '#/components/schemas/PageMeta'
...
    UserRole:
      type: string
      enum: [PATIENT, DOCTOR, ADMIN]
    AppointmentStatus:
      type: string
      enum: [SCHEDULED, CANCELLED, RESCHEDULED, COMPLETED]
...
    UserProfile:
      type: object
      required: [id, name, email, role]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/UserRole'
        clinicId:
          type: string
          format: uuid
        phoneNumber:
          type: string
        dateOfBirth:
          type: string
          format: date
        specialization:
          type: string
        licenseNumber:
          type: string
        enabled:
          type: boolean
          description: Whether the account is active. Disabled accounts cannot authenticate.
        password:
          type: string
          writeOnly: true
...
    Appointment:
      type: object
      required: [id, patientId, doctorId, dateTime, status, duration]
      properties:
        id:
          type: string
          format: uuid
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        dateTime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        duration:
          type: integer
          description: Duration in minutes
        reason:
          type: string
...
    VisitHistory:
      type: object
      required: [patientId, appointments]
      properties:
        patientId:
          type: string
          format: uuid
        appointments:
          type: array
          items:
            $ref: '#/components/schemas/Appointment'
...