# services/springboot/generated/src/main/java/org/openapitools/api/PatientsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.springframework.lang.Nullable;
import org.openapitools.model.PaginatedAppointmentResponse;
import org.openapitools.model.ProblemDetail;
import java.util.UUID;
import org.openapitools.model.UserProfile;
import org.openapitools.model.VisitHistory;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Tag(name = "Patients", description = "Patient profile and history operations")
public interface PatientsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_PATIENT_BY_ID = "/patients/{patientId}";
    /**
     * GET /patients/{patientId} : Get patient profile
     *
     * @param patientId  (required)
     * @return Patient profile (status code 200)
     *         or The request is malformed or fails validation. (status code 400)
     *         or Authentication is required or has failed. (status code 401)
     *         or The caller is authenticated but not allowed to access the resource. (status code 403)
     *         or The requested resource does not exist. (status code 404)
     *         or An unexpected error occurred while processing the request. (status code 500)
     */
    @Operation(
        operationId = "getPatientById",
        summary = "Get patient profile",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Patient profile", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = UserProfile.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = UserProfile.class))
            }),
            @ApiResponse(responseCode = "400", description = "The request is malformed or fails validation.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication is required or has failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "403", description = "The caller is authenticated but not allowed to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource does not exist.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred while processing the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = PatientsApi.PATH_GET_PATIENT_BY_ID,
        produces = { "application/json", "application/problem+json" }
    )
    default ResponseEntity<UserProfile> getPatientById(
        @Parameter(name = "patientId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("patientId") UUID patientId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"clinicId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"password\" : \"password\", \"role\" : \"PATIENT\", \"phoneNumber\" : \"phoneNumber\", \"name\" : \"name\", \"specialization\" : \"specialization\", \"dateOfBirth\" : \"2000-01-23\", \"licenseNumber\" : \"licenseNumber\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"email\" : \"email\", \"enabled\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PATIENT_VISIT_HISTORY = "/patients/{patientId}/visit-history";
    /**
     * GET /patients/{patientId}/visit-history : View a patient&#39;s visit history
     *
     * @param patientId  (required)
     * @return Visit history summary (status code 200)
     *         or The request is malformed or fails validation. (status code 400)
     *         or Authentication is required or has failed. (status code 401)
     *         or The caller is authenticated but not allowed to access the resource. (status code 403)
     *         or The requested resource does not exist. (status code 404)
     *         or An unexpected error occurred while processing the request. (status code 500)
     */
    @Operation(
        operationId = "getPatientVisitHistory",
        summary = "View a patient's visit history",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Visit history summary", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = VisitHistory.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = VisitHistory.class))
            }),
            @ApiResponse(responseCode = "400", description = "The request is malformed or fails validation.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication is required or has failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "403", description = "The caller is authenticated but not allowed to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource does not exist.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred while processing the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = PatientsApi.PATH_GET_PATIENT_VISIT_HISTORY,
        produces = { "application/json", "application/problem+json" }
    )
    default ResponseEntity<VisitHistory> getPatientVisitHistory(
        @Parameter(name = "patientId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("patientId") UUID patientId
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"appointments\" : [ { \"dateTime\" : \"2000-01-23T04:56:07.000+00:00\", \"duration\" : 0, \"reason\" : \"reason\", \"patientId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"doctorId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"SCHEDULED\" }, { \"dateTime\" : \"2000-01-23T04:56:07.000+00:00\", \"duration\" : 0, \"reason\" : \"reason\", \"patientId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"doctorId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"SCHEDULED\" } ], \"patientId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_PATIENT_APPOINTMENTS = "/patients/{patientId}/appointments";
    /**
     * GET /patients/{patientId}/appointments : List a patient&#39;s appointments (paginated)
     *
     * @param patientId  (required)
     * @param page Zero-based page index. (optional, default to 0)
     * @param size Number of items per page. (optional, default to 20)
     * @return Paginated patient appointments (status code 200)
     *         or The request is malformed or fails validation. (status code 400)
     *         or Authentication is required or has failed. (status code 401)
     *         or The caller is authenticated but not allowed to access the resource. (status code 403)
     *         or The requested resource does not exist. (status code 404)
     *         or An unexpected error occurred while processing the request. (status code 500)
     */
    @Operation(
        operationId = "listPatientAppointments",
        summary = "List a patient's appointments (paginated)",
        tags = { "Patients" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Paginated patient appointments", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PaginatedAppointmentResponse.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = PaginatedAppointmentResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "The request is malformed or fails validation.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "401", description = "Authentication is required or has failed.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "403", description = "The caller is authenticated but not allowed to access the resource.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "404", description = "The requested resource does not exist.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            }),
            @ApiResponse(responseCode = "500", description = "An unexpected error occurred while processing the request.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProblemDetail.class)),
                @Content(mediaType = "application/problem+json", schema = @Schema(implementation = ProblemDetail.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = PatientsApi.PATH_LIST_PATIENT_APPOINTMENTS,
        produces = { "application/json", "application/problem+json" }
    )
    default ResponseEntity<PaginatedAppointmentResponse> listPatientAppointments(
        @Parameter(name = "patientId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("patientId") UUID patientId,
        @Min(value = 0) @Parameter(name = "page", description = "Zero-based page index.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "Number of items per page.", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"page\" : { \"size\" : 1, \"totalPages\" : 0, \"page\" : 0, \"totalElements\" : 0 }, \"content\" : [ { \"dateTime\" : \"2000-01-23T04:56:07.000+00:00\", \"duration\" : 0, \"reason\" : \"reason\", \"patientId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"doctorId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"SCHEDULED\" }, { \"dateTime\" : \"2000-01-23T04:56:07.000+00:00\", \"duration\" : 0, \"reason\" : \"reason\", \"patientId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"doctorId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"status\" : \"SCHEDULED\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/problem+json"))) {
                    String exampleString = "Custom MIME type example not yet supported: application/problem+json";
                    ApiUtil.setExampleResponse(request, "application/problem+json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}